Refactor policy steps
